requirements.txt
wikipedia==1.4.0
requests-html==0.10.0
requests==2.28.2
torch==1.13.1
transformers==4.26.1
fastapi==0.94.1
uvicorn==0.21.0
spacy==3.5.1
google-cloud-dialogflow==2.21.0
google-api-core==1.34.0